Fix Kotlin type mismatch in NDEF TNF dispatch
when(record.tnf.toInt()) compared an Int subject against the Short
TNF_* constants, which Kotlin doesn't implicitly widen -- this broke
compilation of the whole nfc package. Switch on the Short tnf directly.

+++ app/src/main/java/com/zephyrcloud/nfckit/nfc/NdefRecordParser.kt
@@ -12,7 +12,7 @@ object NdefRecordParser {
 
     fun parseRecord(record: NdefRecord): ParsedNdefRecord {
         val type = String(record.type, Charsets.US_ASCII)
-        return when (record.tnf.toInt()) {
+        return when (record.tnf) {
             NdefRecord.TNF_WELL_KNOWN -> when {
                 type == String(NdefRecord.RTD_TEXT, Charsets.US_ASCII) -> parseText(record)
                 type == String(NdefRecord.RTD_URI, Charsets.US_ASCII) -> parseUri(record)
